requirements.txt
# Core AutoGen Framework
pyautogen==0.2.9

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Environment Management
python-dotenv==1.0.0

# Output Formatting
tabulate==0.9.0
colorama==0.4.6

# Date/Time Handling
python-dateutil==2.8.2

# Email (for future integration)
email-validator==2.1.0

# Testing
pytest==7.4.3